Updated the InstallShield LE project to not scan shipped assemblies for dependencies at build time. There appears to be a problem with InstallShield 2012 Spring LE which causes it to incorrectly detect a dependency on Visual C++ V11.0 for ARM processors. This could cause installations for previous versions of CodeGen built for Synergy/DE V10 to incorrectly install the ARM version of the C++ V11 runtime if C++ V11 was not already installed on the system.
This could cause a problem if CodeGen for Synergy V10 was incorrectly installed on a Synergy/DE V9 system and then Synergy/DE was upgraded to V10. The Synergy/DE installation would see that a version of C++ V11 was already present (but it was the ARM version), but most Synergy/DE components would fail to execute after installation because of the incorrect C++ runtime.

If you encounter this problem then the correct solution is to first delete MSVCP110.DLL and MSVCR110.DLL from your Windows\System32 folder (or your Windows\SysWOW64 folder if you are on a 64-bit system) and then use the Programs and Features utility in Windows Control Panel to perform a “Repair” operation on the 32-bit Synergy/DE installation.